requirements.txt
requests
pushbullet.py
beautifulsoup4
